+++ CHANGELOG.md
@@ -4,15 +4,25 @@ All notable changes to this project will be documented in this file. The format 
 
 alternate-node-red-installer adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased](https://github.com/TotallyInformation/alternate-node-red-installer/compare/v2.0.1...master)
+## [Unreleased](https://github.com/TotallyInformation/alternate-node-red-installer/compare/v1.0.0...master)
+
+### Breaking
+
+- Node.js v10 is now the minimum version due to dependency module changes.
 
 ### Added
 
 - [CHANGELOG](./CHANGELOG.md) and [TODO](./TODO.md) files.
-- Bump Ora and Yargs dependencies to latest versions.
+
+### Changed
+
+- Bump to v2.0 due to breaking change.
+- Bump dependencies to latest versions.
+- Tidy and improve code.
+- Install now sets correct paths in `./system/node-red-service` and `./data/envfile.ini`.
 
 ----
 
-## [1.0.2](https://github.com/TotallyInformation/alternate-node-red-installer/compare/v2.0.1...v2.0.2) - 
+## [1.0.0](https://github.com/TotallyInformation/alternate-node-red-installer/compare/v0.0.1...v1.0.0) - 
 
 ### Fixed